+++ OrditiApp/src/app/busca/busca.page.html
@@ -4,7 +4,7 @@
       <ion-menu-button></ion-menu-button>
     </ion-buttons>
     <ion-title>
-      Buscar
+      Buscar Região
     </ion-title>
   </ion-toolbar>
   <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
