+++ OrditiApp/src/app/home/home.page.scss
@@ -4,7 +4,6 @@
 }
 
 #logo{
-	width: 45%;
 }
 
 ion-toolbar ion-title{
